refactor (hatchery/VSphere)
